Real Estate Project on Housing in United State
Project: Housing in United State
Data Source: Realtor
Tool: Jupyter Notebook
Approach: EDA

Project Overview:
Objective:
The primary goal of this project is to analyze the real estate market and explore key factors that influence property sale prices. Specifically, the project focuses on:
·      Determining whether property size or/and location has a greater impact on sale prices.
·      Identifying which state has the most expensive real estate market.
·      Investigating the relationship between home size/property size (square footage) and sale price.

Dataset:
The project uses a real estate dataset containing various attributes, including acre lot, price, location (state, city) etc.

Key Analyses:
Influence of Property Size and Location on Sale Price:
An Exploratory Data Analysis (EDA) was used to assess the correlation between property size and sale price, while also accounting for the impact of location on price variations.

State-by-State Market Comparison:
The dataset is analyzed to rank states by average property prices, highlighting the most expensive real estate markets.

Relationship Between Home Size and Price:
Scatter plots, correlation coefficients are applied to explore how changes in home size (area in m2) affect pricing.

Findings:
After performing an EDA, it was seen that there is a weak positive correlation between home size (area in m2) and price for different locations. This simply mean that even though the home size and location is part of the factors that determine the price of a home, there are other factors that are also considered when determining the price of a home. Also, the state with the most expensive real estate market was discovered to be Utah.

Recommendation:
Based on the findings, it's clear that while home size (area in m2) and location play a role in determining property prices, it is not the sole driver, so when purchasing or investing in real estate, factors beyond property size and location should be considered.